Feature: switches for viscous dissipation terms
1. Added switches to consider/neglect viscous dissipation terms from the
energy equation (temperature), thus considering/disregarding the heat
generation due to these mechanisms.

The switches are activated in the "parameters" IOdictionary from file
onstant/constitutiveProperties, under the following boolean flags:

	newtonianDissipationTerm
	viscoelasticDissipationTerm

The first switch acts on the newtonian viscous dissipation term,
whilst the second concearns the viscoelastic extra-terms to the
energy equation. By default, if a flag is not found it is considered
to have a "true" value. A "true" value implies that the respective
dissipation term is considered, a "false" value removes the term from
the computations. Check the channel_noHeatFromViscousDissipation tutorial
for an example.

2. Created solver rheoThemFoamOptimized which contains the rheoThermFoam
code without these switches, for optimized run without the delay
introduced by these features.

3. Deleted solver rheoThermFoamNoDiss as it became redundant.